Conflicts:
	DofusProtocol/D2oClasses/Stump.DofusProtocol.D2oClasses.csproj

+++ DofusProtocol/Messages/Messages/game/context/fight/GameFightJoinMessage.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 01/04/2015 11:54:11
+// Generated on 08/04/2015 13:24:55
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -21,14 +21,14 @@ namespace Stump.DofusProtocol.Messages
         public bool canBeCancelled;
         public bool canSayReady;
         public bool isFightStarted;
-        public int timeMaxBeforeFightStart;
+        public short timeMaxBeforeFightStart;
         public sbyte fightType;
         
         public GameFightJoinMessage()
         {
         }
         
-        public GameFightJoinMessage(bool canBeCancelled, bool canSayReady, bool isFightStarted, int timeMaxBeforeFightStart, sbyte fightType)
+        public GameFightJoinMessage(bool canBeCancelled, bool canSayReady, bool isFightStarted, short timeMaxBeforeFightStart, sbyte fightType)
         {
             this.canBeCancelled = canBeCancelled;
             this.canSayReady = canSayReady;
@@ -44,7 +44,7 @@ namespace Stump.DofusProtocol.Messages
             flag1 = BooleanByteWrapper.SetFlag(flag1, 1, canSayReady);
             flag1 = BooleanByteWrapper.SetFlag(flag1, 2, isFightStarted);
             writer.WriteByte(flag1);
-            writer.WriteInt(timeMaxBeforeFightStart);
+            writer.WriteShort(timeMaxBeforeFightStart);
             writer.WriteSByte(fightType);
         }
         
@@ -54,7 +54,7 @@ namespace Stump.DofusProtocol.Messages
             canBeCancelled = BooleanByteWrapper.GetFlag(flag1, 0);
             canSayReady = BooleanByteWrapper.GetFlag(flag1, 1);
             isFightStarted = BooleanByteWrapper.GetFlag(flag1, 2);
-            timeMaxBeforeFightStart = reader.ReadInt();
+            timeMaxBeforeFightStart = reader.ReadShort();
             if (timeMaxBeforeFightStart < 0)
                 throw new Exception("Forbidden value on timeMaxBeforeFightStart = " + timeMaxBeforeFightStart + ", it doesn't respect the following condition : timeMaxBeforeFightStart < 0");
             fightType = reader.ReadSByte();
